+++ tests/testthat/test_ncdf.R
@@ -63,8 +63,7 @@ test_that("curvilinear", {
   warn <- capture_warnings(out <-read_ncdf(f, curvilinear = c("lon", "lat")))
   
   expect_match(warn[1], "Could not parse expression:.*Returning as a single symbolic unit()")
-  expect_match(warn[2], "ignoring unrecognized unit: kg m\\^-2")
-  expect_match(warn[3], "bounds for time seem to be reversed; reverting them")
+  expect_match(warn[2], "bounds for time seem to be reversed; reverting them")
   
   st_dim <- st_dimensions(out)